+++ packages/product/CHANGELOG.md
@@ -14,4 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.6.0] - 2026-06-20
+
+### Added
+
+- Use micro UUID document IDs for generated PRD artifact filenames instead of sequence numbers.
+
 - Plugin extraction and population (in progress)